+++ tests/test_visitors/test_ast/test_naming/test_variable_usages.py
@@ -9,6 +9,10 @@ from wemake_python_styleguide.visitors.ast.naming import (
     WrongNameVisitor,
 )
 
+import_name = 'import {0}'
+from_import_module = 'from {0} import some'
+from_import_name = 'from some import {0}'
+
 calling_function = 'print({0})'
 called_function = '{0}()'
 calling_method = 'instance.call({0})'
@@ -33,8 +37,19 @@ def function():
 """
 
 
-@pytest.mark.parametrize('bad_name', VARIABLE_NAMES_BLACKLIST)
+@pytest.mark.parametrize('bad_name', [
+    *VARIABLE_NAMES_BLACKLIST,
+    *string.ascii_letters,
+    '__Class_private',
+    'number_prefix_10',
+    'some__underscores',
+    'camelCase',
+    'UPPER_case',
+])
 @pytest.mark.parametrize('code', [
+    import_name,
+    from_import_module,
+    from_import_name,
     calling_function,
     called_function,
     calling_method,
@@ -47,66 +62,6 @@ def function():
     yielding_variable,
 ])
 def test_wrong_variable_names(
-    assert_errors,
-    parse_ast_tree,
-    bad_name,
-    code,
-    default_options,
-):
-    """Testing that any variable can used without raising violations."""
-    tree = parse_ast_tree(code.format(bad_name))
-
-    visitor = WrongNameVisitor(default_options, tree=tree)
-    visitor.run()
-
-    assert_errors(visitor, [])
-
-
-@pytest.mark.parametrize('bad_name', string.ascii_letters)
-@pytest.mark.parametrize('code', [
-    calling_function,
-    called_function,
-    calling_method,
-    called_method,
-    accessing_attribute,
-    accessed_attribute,
-    raising_variable,
-    returning_variable,
-    awaiting_variable,
-    yielding_variable,
-])
-def test_short_variable_names(
-    assert_errors,
-    parse_ast_tree,
-    bad_name,
-    code,
-    default_options,
-):
-    """Testing that any variable can used without raising violations."""
-    tree = parse_ast_tree(code.format(bad_name))
-
-    visitor = WrongNameVisitor(default_options, tree=tree)
-    visitor.run()
-
-    assert_errors(visitor, [])
-
-
-@pytest.mark.parametrize('bad_name', [
-    '__Class_private',
-])
-@pytest.mark.parametrize('code', [
-    calling_function,
-    called_function,
-    calling_method,
-    called_method,
-    accessing_attribute,
-    accessed_attribute,
-    raising_variable,
-    returning_variable,
-    awaiting_variable,
-    yielding_variable,
-])
-def test_private_variable_names(
     assert_errors,
     parse_ast_tree,
     bad_name,